+++ integration_tests/ethereum/hardhat.config.ts
@@ -1,75 +1,85 @@
+import '@nomiclabs/hardhat-ethers';
 import '@nomiclabs/hardhat-waffle';
-import {task} from 'hardhat/config';
+import {task} from "hardhat/config";
 import * as constants from "./addresses";
+
+export async function deployContracts(
+    ethers: any,
+    gravityId: string = "foo",
+    valAddresses: string[],
+    powers: number[],
+    powerThreshold: number
+) {
+    console.log(`creating gravity contract`)
+    const Gravity = await ethers.getContractFactory("Gravity");
+    console.log(`creating checkpoint`)
+    // const checkpoint = makeCheckpoint(ethers, valAddresses, powers, 0, gravityId);
+    console.log(`deploying gravity contract`)
+    const gravity = (await Gravity.deploy(
+        ethers.utils.formatBytes32String(gravityId),
+        powerThreshold,
+        valAddresses,
+        powers
+    ));
+
+    await gravity.deployed();
+    console.log(`gravity contract deployed at - ${gravity.address}`)
+
+    return { gravity };
+}
 
 task(
     'integration_test_setup',
     'Sets up contracts for the integration test',
     async (args, hre) => {
 
+        let powers: number[] = [1073741824,1073741824,1073741824,1073741824];
+        let powerThreshold: number = 6666;
+        let {gravity} = await deployContracts(hre.ethers,"gravitytest", constants.VALIDATORS, powers, powerThreshold);
 
-        // console.log('retrieving gravity contract');
-        // const gravitySigner = await hre.ethers.getSigner(ADDRESSES.GRAVITY_OWNER)
-        // // Attach Gravity contract
-        // const Gravity = hre.ethers.getContractAt(
-        //     'Gravity',
-        //     ADDRESSES.GRAVITY,
-        //     gravitySigner
-        // );
-        // const gravity = await Gravity;
-        // console.log(`attached to gravity: ${gravity}`)
+        console.log('taking over cellar owner');
+        // Take over the cellar owner so we can transfer ownership
+        await hre.network.provider.request({
+            method: 'hardhat_impersonateAccount',
+            params: [constants.CELLAR_OWNER],
+        });
+        const cellarSigner = await hre.ethers.getSigner(constants.CELLAR_OWNER);
+        const Cellar = hre.ethers.getContractAt(
+            'CellarPoolShare',
+            constants.CELLAR,
+            cellarSigner,
+        );
+        const cellar = await Cellar;
 
-        // console.log('taking over cellar owner');
-        // // Take over the cellar owner so we can transfer ownership
-        // await hre.network.provider.request({
-        //     method: 'hardhat_impersonateAccount',
-        //     params: [ADDRESSES.CELLAR_OWNER],
-        // });
-
-        // console.log('getting cellar signer');
-        // const cellarSigner = await hre.ethers.getSigner(ADDRESSES.CELLAR_OWNER);
-        //
-        // console.log('retrieving cellar contract');
-        // // Transfer ownership to gravity
-        // const Cellar = hre.ethers.getContractAt(
-        //     'CellarPoolShare',
-        //     ADDRESSES.CELLAR,
-        //     cellarSigner,
-        // );
-        // const cellar = await Cellar;
-        //
-        // let { hash } = await cellar.transferOwnership(ADDRESSES.GRAVITY, {
-        //     gasPrice: hre.ethers.BigNumber.from('99916001694'),
-        // });
-        //
-        // // Send ETH to needed accounts
-        //
-        // console.log(
-        //     `Cellar contract at ${ADDRESSES.CELLAR} is now owned by Gravity contract at ${ADDRESSES.GRAVITY}`,
-        // );
-        // console.log(`Tx hash: ${hash}`);
-        // console.log('='.repeat(80));
+        let { hash } = await cellar.transferOwnership(gravity.address, {
+            gasPrice: hre.ethers.BigNumber.from('99916001694'),
+        });
+        console.log(
+            `Cellar contract at ${cellar.address} is now owned by Gravity contract at ${gravity.address}`,
+        );
 
         // Take over vitalik.eth
-        // await hre.network.provider.request({
-        //     method: 'hardhat_impersonateAccount',
-        //     params: [constants.WHALE],
-        // });
-        //
-        // // Send ETH to needed parties
-        // const whaleSigner = await hre.ethers.getSigner(constants.WHALE);
-        //
-        // for (let addr of constants.VALIDATORS) {
-        //     await whaleSigner.sendTransaction({
-        //         to: addr,
-        //         value: hre.ethers.utils.parseEther('100'),
-        //     });
-        // }
+        console.log("taking over whale account")
+        await hre.network.provider.request({
+            method: 'hardhat_impersonateAccount',
+            params: [constants.WHALE],
+        });
 
-        // start the ethereum node after all setup is complete
-        // await hre.run('node');
-    },
-);
+        // Send ETH to needed parties
+        const whaleSigner = await hre.ethers.getSigner(constants.WHALE);
+
+        for (let addr of constants.VALIDATORS) {
+            console.log(`sending 100 eth to: ${addr}`);
+            await whaleSigner.sendTransaction({
+                to: addr,
+                value: hre.ethers.utils.parseEther('100'),
+            });
+        }
+
+        console.log("node is listening")
+        await hre.run('node');
+    });
+
 
 /**
  * @type import('hardhat/config').HardhatUserConfig
@@ -116,7 +126,6 @@ module.exports = {
             },
         ],
     },
-    // TODO: add forking configuration
     typechain: {
         outDir: 'typechain',
         target: 'ethers-v5',
