+++ integration_tests/ethereum/package.json
@@ -4,9 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "start": "npx hardhat integration_test_setup --verbose",
-    "deploy": "npx hardhat run scripts/deploy.ts --network hardhat",
     "compile": "npx hardhat compile"
   },
   "keywords": [],
