+++ integration_tests/setup_test.go
@@ -61,7 +61,7 @@ var (
 	stakeAmount, _  = sdk.NewIntFromString("100000000000")
 	stakeAmountCoin = sdk.NewCoin(bondDenom, stakeAmount)
 	hardhatCellar   = common.HexToAddress("0x6ea5992aB4A78D5720bD12A089D13c073d04B55d")
-	gravityContract = common.HexToAddress("0xFbB0BCfed0c82043A7d5387C35Ad8450b44A4cde")
+	gravityContract = common.HexToAddress("0x04C89607413713Ec9775E14b954286519d836FEf")
 )
 
 type IntegrationTestSuite struct {
@@ -453,7 +453,6 @@ func (s *IntegrationTestSuite) runEthContainer() {
 			"8545/tcp": {{HostIP: "", HostPort: "8545"}},
 		},
 		ExposedPorts: []string{"8545/tcp"},
-		Cmd: []string{"yarn", "run", "deploy"},
 	}
 
 	s.ethResource, err = s.dockerPool.RunWithOptions(
@@ -471,12 +470,18 @@ func (s *IntegrationTestSuite) runEthContainer() {
 			ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
 			defer cancel()
 
-			balance, err := ethClient.BalanceAt(ctx, common.HexToAddress("0xd312f0f1B39D54Db2829537595fC1167B14d4b34"), nil)
+			balance, err := ethClient.BalanceAt(ctx, common.HexToAddress(s.chain.validators[0].ethereumKey.address), nil)
 			if err != nil {
+				s.T().Logf("error querying balance: %e", err)
 				return false
 			}
 
-			if balance == nil || (balance.Cmp(big.NewInt(0)) == 0) {
+			if balance == nil {
+				s.T().Logf("balance for first validator is nil")
+			}
+
+			if balance.Cmp(big.NewInt(0)) == 0 {
+				s.T().Logf("balance for first validator is %s", balance.String())
 				return false
 			}
 
@@ -499,7 +504,7 @@ func (s *IntegrationTestSuite) runEthContainer() {
 	s.Require().Eventuallyf(func() bool {
 
 		for _, s := range strings.Split(ethereumLogOutput.String(), "\n") {
-			if strings.HasPrefix(s, "Gravity deployed at Address") {
+			if strings.HasPrefix(s, "gravity contract deployed at") {
 				strSpl := strings.Split(s, "-")
 				gravityContract = common.HexToAddress(strings.ReplaceAll(strSpl[1], " ", ""))
 				return true
@@ -507,7 +512,7 @@ func (s *IntegrationTestSuite) runEthContainer() {
 		}
 		return false
 	}, time.Minute * 5, time.Second * 10, "unable to retrieve gravity address from logs")
-
+	s.T().Logf("gravity contrained deployed at %s", gravityContract.String())
 
 	s.T().Logf("started Ethereum container: %s", s.ethResource.Container.ID)
 }
